Rename project and namespaces to Auth0.ManagementApi.Client
