+++ Auth0.Net.sln
@@ -12,7 +12,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{C9550C3F-90E
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Auth0.Core", "src\Auth0.Core\Auth0.Core.csproj", "{C5395825-EF6D-438E-9E22-95975F1276E8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Auth0.Api.Management", "src\Auth0.Api.Management\Auth0.Api.Management.csproj", "{01898E06-FA70-42B8-AE35-28D2BACDFBD7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Auth0.ManagementApi.Client", "src\Auth0.ManagementApi.Client\Auth0.ManagementApi.Client.csproj", "{01898E06-FA70-42B8-AE35-28D2BACDFBD7}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{0EBAD055-E431-443B-8212-411D1AFE3796}"
 EndProject
