--- src/Auth0.Api.Management/Clients/ClientBase.cs
@@ -1,12 +1,0 @@
-namespace Auth0.Api.Management.Clients
-{
-    public class ClientBase
-    {
-        protected IApiConnection Connection { get; }
-
-        public ClientBase(IApiConnection connection)
-        {
-            Connection = connection;
-        }
-    }
-}
