+++ src/Auth0.ManagementApi.Client/Clients/ClientBase.cs
@@ -1,0 +1,12 @@
+namespace Auth0.ManagementApi.Client.Clients
+{
+    public class ClientBase
+    {
+        protected IApiConnection Connection { get; }
+
+        public ClientBase(IApiConnection connection)
+        {
+            Connection = connection;
+        }
+    }
+}
