+++ src/Auth0.ManagementApi.Client/Clients/EmailProviderClient.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Auth0.Core.Models;
 using PortableRest;
 
-namespace Auth0.Api.Management.Clients
+namespace Auth0.ManagementApi.Client.Clients
 {
     public class EmailProviderClient : ClientBase, IEmailProviderClient
     {
