+++ src/Auth0.ManagementApi.Client/Clients/IStatsClient.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Auth0.Core.Models;
 
-namespace Auth0.Api.Management.Clients
+namespace Auth0.ManagementApi.Client.Clients
 {
     public interface IStatsClient
     {
