+++ src/Auth0.ManagementApi.Client/Clients/StatsClient.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Auth0.Core.Models;
 
-namespace Auth0.Api.Management.Clients
+namespace Auth0.ManagementApi.Client.Clients
 {
     public class StatsClient :  ClientBase, IStatsClient
     {
