+++ src/Auth0.ManagementApi.Client/Properties/AssemblyInfo.cs
