+++ src/Auth0.ManagementApi.Client/app.config
