+++ src/Auth0.ManagementApi.Client/packages.config
